feat: update schema social-link
