+++ lib/queries/social-network.js
@@ -2,10 +2,11 @@ import { groq } from 'next-sanity';
 
 const socialItem = `
 _id,
-social_name,
+name,
+icon,
 social_item_link
 `;
 
-export const socialNetworksListQuery = groq`*[_type == "social_link"][]{${socialItem}}`;
+export const socialNetworksListQuery = groq`*[_type == "social_link_list"][]{${socialItem}}`;
 
 export default socialNetworksListQuery;
