+++ studio/schemas/schema.js
@@ -10,11 +10,11 @@ import schemaTypes from 'all:part:@sanity/base/schema-type';
 // import blockContent from './blockContent'
 import post from './post';
 import author from './author';
-import socialLink from './social-link';
 import donationV1 from './donation-v1';
 import donationV2 from './donation-v2';
 import speakers from './speakers';
 import community from './community';
+import socialLinkList from './social-link-list';
 
 // Then we give our schema to the builder and provide the result to Sanity
 export default createSchema({
@@ -26,7 +26,7 @@ export default createSchema({
     /* Your types here! */
     post,
     author,
-    socialLink,
+    socialLinkList,
     donationV1,
     donationV2,
     speakers,
