+++ studio/schemas/social-link-list.js
@@ -1,16 +1,22 @@
 export default {
-  name: 'social_link',
-  title: 'SocialLink',
+  name: 'social_link_list',
+  title: 'SocialLinkList',
   type: 'document',
   fields: [
     {
       title: 'Social name',
-      name: 'social_name',
+      name: 'name',
       type: 'string',
       validation: Rule => Rule.required()
     },
     {
-      title: 'Social link',
+      title: 'Social icon',
+      name: 'icon',
+      type: 'image',
+      validation: Rule => Rule.required()
+    },
+    {
+      title: 'Social link list',
       name: 'social_item_link',
       type: 'array',
       of: [
@@ -19,7 +25,6 @@ export default {
           fields: [
             { name: 'name', type: 'string', title: 'Name link' },
             { name: 'link', type: 'url', title: 'URL' },
-            { name: 'country', type: 'string', title: 'Country' }
           ]
         }
       ],
